Added more fixes
