+++ README.md
@@ -7,8 +7,8 @@
 
  **Hirola** is an opinionated web framework for that is focused on simplicity and predicatability.
 
-#### Example
-lib.rs
+Here is a simple example: 
+
 ```rust
 use hirola::prelude::*;
 use std::sync::Arc;
@@ -16,6 +16,8 @@ use std::sync::Arc;
 #[derive(Default)]
 struct Count {
     counter: Reactive<i32>,
+    display: Reactive<bool>,
+    non_reactive: String
 }
 
 impl Count {
@@ -37,26 +39,25 @@ struct Counter;
 impl Component<Option<i32>, Count> for Counter {
     fn render(&mut self, state: &Arc<Count>) -> Dom {
         render! {
-            <div data-x-show={true} class="flex w-full h-full">
-                <div class="custom-number-input h-10 w-32">
-                    <label for="custom-input-number" class="w-full text-gray-700 text-sm font-semibold">"Counter Input"
-                    </label>
-                    <div class="flex flex-row h-10 w-full rounded-lg relative bg-transparent mt-1">
+            <div data-x-show={&state.display} class="flex w-full h-full">
+                <div class="h-10 w-32">
+                    <div class="flex flex-row h-10">
                         <button 
-                            data-action="decrement" 
-                            onclick={ |_event| { &state.decrement() }}
-                            class="bg-gray-300 text-gray-600 h-full w-20 rounded-l cursor-pointer outline-none">
-                            <span class="m-auto text-2xl font-thin">"−"</span>
+                            // data-x-transition={&state.animation.button} Coming soon
+                            // data-action="decrement" Comments allowed 
+                            onclick={ |_event|  &state.decrement(); }
+                            class="bg-gray-300">
+                            "+"
                         </button>
                         <input 
-                            class="outline-none text-center w-full bg-gray-300 font-semibold text-md flex items-center text-gray-700 outline-none" data-name="custom-input-number" 
-                            value={state.counter}
+                            class="outline-none text-center"
+                            value={&state.counter}
                         />
                         <button 
                             data-action="increment"
-                            onclick={ |_event| { &state.increment() }}
-                            class="bg-gray-300 text-gray-600  h-full w-20 rounded-r cursor-pointer">
-                            <span class="m-auto text-2xl font-thin">"+"</span>
+                            onclick={ |_event| &state.increment(); }
+                            class="bg-gray-300">
+                            "+"
                         </button>
                     </div>
                 </div>
@@ -64,119 +65,26 @@ impl Component<Option<i32>, Count> for Counter {
         }
     }
 }
-
-#[wasm_bindgen(start)]
-pub fn main() -> Result<(), JsValue> {
-    #[cfg(debug_assertions)]
-    console_error_panic_hook::set_once();
-    hirola::mount(&mut Counter, None);
-    Ok(())
-}
-
 ```
 
-public/index.html
-```html
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Hirola Counter</title>
-    <link href="https://unpkg.com/tailwindcss@^2/dist/tailwind.min.css" rel="stylesheet">
-  </head>
-  <body>
-    <script src="js/index.js"></script>
-  </body>
-</html>
-```
-In Cargo.toml, specify the crate-type to be `cdylib`
-```toml
+### Goals
+- [x] Write code that is declarative and easy to follow.
+- [x] Follow Alpine-ish kind of Reactive and declarative style.
+- [ ] Extensible for other gui. Since the core principles dont care about UI.
+- [ ] Async Handling and server-side integration.
 
-[package]
-name = "counter"
-version = "0.1.0"
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib"]
-
-
-[dependencies]
-hirola = "0.1.0"
-console_error_panic_hook = "0.1"
-log = "0.4"
-console_log = "0.2"
-```
-
-package.json
-```json
-{
-    "private": true,
-    "author": "@geofmureithi",
-    "name": "counter",
-    "version": "0.1.0",
-    "scripts": {
-      "build": "rimraf dist/js && rollup --config",
-      "start": "rimraf dist/js && rollup --config --watch"
-    },
-    "devDependencies": {
-      "@wasm-tool/rollup-plugin-rust": "^1.0.0",
-      "rimraf": "^3.0.2",
-      "rollup": "^1.31.0",
-      "rollup-plugin-livereload": "^1.2.0",
-      "rollup-plugin-serve": "^1.0.1"
-    }
-  }
-  
-```
-rollup.config.js
-```js
-import rust from "@wasm-tool/rollup-plugin-rust";
-import serve from "rollup-plugin-serve";
-import livereload from "rollup-plugin-livereload";
-
-const is_watch = !!process.env.ROLLUP_WATCH;
-
-export default {
-    input: {
-        index: "./Cargo.toml",
-    },
-    output: {
-        dir: "public/js",
-        format: "iife",
-        sourcemap: true,
-    },
-    plugins: [
-        rust({
-            serverPath: "js/",
-            debug: false,
-        }),
-
-        is_watch && serve({
-            contentBase: "public",
-            open: true,
-        }),
-
-        is_watch && livereload("public"),
-    ],
-};
-```
-
-
-
-Start using
-```sh
-$> yarn start
-```
-
-Build using
-```sh
-$> yarn build
-```
-
+### Inspiration
+I was inspired by alot of We frameworks out there. You can see influences of React in the `Component`, `State` and `Props` traits.
+You can also see inspirations from `yew` and other similar frameworks to provide a seemless macro.
+I also want to capture the declarativeness and reactiveness seen in alpine.js. Expect to see more of that direction, eg. with transitions.
 
 #### Demo examples
 - counter ToDO
+
+PS
+> The above example doesnt work. You can try the `develop` branch though.
+
+> This API will certainly change.
 
 #### Prerequisite:
 
@@ -184,7 +92,8 @@ You need need to have `rust` and `cargo` installed.
 For the counter example, you need `node.js` and `npm` installed.
 
 Note:
-> The above example doesnt work yet.. The counter example in the 
+> The above example doesnt work yet.. The counter example in the develop branch 
+
 > `cargo web` doesnt work possibly because of wasm-bindgen
