+++ crates/hirola-core/src/events.rs
@@ -1,157 +1,9 @@
 // from dominator events
 
 use dominator::traits::StaticEvent;
-use std::fmt::{Display, Error, Formatter};
-use typed_html::events::EventHandler;
 use wasm_bindgen::JsCast;
 use web_sys::{EventTarget, HtmlInputElement, HtmlTextAreaElement};
 
-use crate::node::Node;
-
-macro_rules! declare_events {
-    ($($name:ident : $type:ty ,)*) => {
-        /// Container type for DOM events.
-        pub struct Events {
-            $(
-                pub $name: Option<Box<dyn EventHandler<Node, $type> + Send>>,
-            )*
-        }
-
-        impl Default for Events {
-            fn default() -> Self {
-                Events {
-                    $(
-                        $name: None,
-                    )*
-                }
-            }
-        }
-
-        /// Iterate over the defined events on a DOM object.
-        #[macro_export]
-        macro_rules! for_events {
-            ($event:ident in $events:expr => $body:block) => {
-                $(
-                    if let Some(ref mut $event) = $events.$name $body
-                )*
-            }
-        }
-    }
-}
-
-// TODO? these are all the "on*" attributes defined in the HTML5 standard, with
-// the ones I've been unable to match to Kutuish event types commented out.
-//
-// This needs review.
-
-declare_events! {
-    //abort: ResourceAbortEvent,
-    // autocomplete: Event,
-    // autocompleteerror: Event,
-    blur: Blur,
-    // cancel: Event,
-    // canplay: Event,
-    // canplaythrough: Event,
-    change: Change,
-    click: Click,    // close: Event,
-    contextmenu: ContextMenu,
-    // cuechange: Event,
-    dblclick: DoubleClick,
-    drag: Drag,
-    dragend: DragEnd,
-    dragenter: DragEnter,
-    //dragexit: DragExit,
-    dragleave: DragLeave,
-    dragover: DragOver,
-    dragstart: DragStart,
-    drop: Drop,
-    // durationchange: Event,
-    // emptied: Event,
-    // ended: Event,
-    //error: ResourceErrorEvent,
-    focus: Focus,
-    input: Input,
-    // invalid: Event,
-    keydown: KeyDown,
-    //keypress: KeyPress,
-    keyup: KeyUp,
-    //load: ResourceLoadEvent,
-    // loadeddata: Event,
-    // loadedmetadata: Event,
-    //loadstart: LoadStartEvent,
-    mousedown: MouseDown,
-    mouseenter: MouseEnter,
-    mouseleave: MouseLeave,
-    mousemove: MouseMove,
-    // mouseout: MouseOut,
-    // mouseover: MouseOver,
-    mouseup: MouseUp,
-    // mousewheel: MouseWheelEvent,
-    // pause: Event,
-    // play: Event,
-    // playing: Event,
-    // progress: ProgressEvent,
-    // ratechange: Event,
-    // reset: Event,
-    resize: Resize,
-    scroll: Scroll,
-    // seeked: Event,
-    // seeking: Event,
-    // select: Event,
-    // show: Event,
-    // sort: Event,
-    // stalled: Event,
-    // submit: Submit,
-    // suspend: Event,
-    // timeupdate: Event,
-    // toggle: Event,
-    // volumechange: Event,
-    // waiting: Event,
-}
-
-impl Display for Events {
-    fn fmt(&self, _f: &mut Formatter) -> Result<(), Error> {
-        Ok(())
-    }
-}
-
-// /// Wrapper type for closures as event handlers.
-// pub struct EFn<F, E>(Option<F>, PhantomData<E>);
-
-// impl<F, E> EFn<F, E>
-// where
-//     F: FnMut(E) + 'static + Send,
-// {
-//     pub fn new(f: F) -> Self {
-//         EFn(Some(f), PhantomData)
-//     }
-// }
-
-// impl<F, E> From<F> for Box<dyn EventHandler<Kutuish, E> + Send>
-// where
-//     F: FnMut(E) + 'static + Send,
-//     E: StaticEvent + 'static + Send,
-// {
-//     fn from(f: F) -> Self {
-//         Box::new(EFn::new(f))
-//     }
-// }
-
-// impl<F, E> EventHandler<Kutuish, E> for EFn<F, E>
-// where
-//     F: FnMut(E) + 'static + Send,
-//     E: StaticEvent + std::marker::Send + 'static,
-// {
-//     fn attach(&mut self, target: &mut <Kutuish as OutputType>::EventTarget) {
-//         let mut handler = self.0.take().unwrap();
-//         (target.event)(&mut handler);
-
-//     }
-
-//     fn render(&self) -> Option<String> {
-//         None
-//     }
-// }
 
 macro_rules! make_event {
     ($name:ident, $type:literal => $event:path) => {
